Rename for improved coherence: Select.js to KeyValueSelector.js, SelectOptions.js to KeyValueSelector_Basic.js, UiLanguageSelector.js to LanguageSelector.js, SelectOptions.sc.js to SelectorStyling.sc.js
