+++ src/components/SelectorStyling.sc.js
