+++ src/pages/LanguagePreferences.js
@@ -12,7 +12,7 @@ import Main from "./info_page_shared/Main";
 import Form from "./info_page_shared/Form";
 import FullWidthErrorMsg from "./info_page_shared/FullWidthErrorMsg";
 import FormSection from "./info_page_shared/FormSection";
-import SelectOptions from "../components/SelectOptions";
+import KeyValueSelector_Basic from "../components/KeyValueSelector_Basic";
 import ButtonContainer from "./info_page_shared/ButtonContainer";
 import Button from "./info_page_shared/Button";
 import ArrowForwardRoundedIcon from "@mui/icons-material/ArrowForwardRounded";
@@ -102,7 +102,7 @@ export default function LanguagePreferences({ api }) {
             <FullWidthErrorMsg>{errorMessage}</FullWidthErrorMsg>
           )}
           <FormSection>
-            <SelectOptions
+            <KeyValueSelector_Basic
               value={learned_language_on_register}
               label={strings.learnedLanguage}
               placeholder={strings.learnedLanguagePlaceholder}
@@ -113,7 +113,7 @@ export default function LanguagePreferences({ api }) {
               onChangeHandler={handleLearned_language_on_register}
             />
 
-            <SelectOptions
+            <KeyValueSelector_Basic
               value={learned_cefr_level_on_register}
               label={strings.levelOfLearnedLanguage}
               placeholder={strings.levelOfLearnedLanguagePlaceholder}
@@ -124,7 +124,7 @@ export default function LanguagePreferences({ api }) {
               onChangeHandler={handleLearned_cefr_level_on_register}
             />
 
-            <SelectOptions
+            <KeyValueSelector_Basic
               value={native_language_on_register}
               label={strings.baseLanguage}
               placeholder={strings.baseLanguagePlaceholder}
